requirements.txt
Django==1.10.1
django-crispy-forms==1.6.0
